Refactor LDAP  and CLI product tests
